move the dailygoal inside the challenge folder
