+++ app/src/androidTest/java/com/epfl/drawyourpath/challenge/StatisticsTest.kt
@@ -1,7 +1,7 @@
 package com.epfl.drawyourpath.challenge
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.`is`
 import org.junit.Test
