+++ app/src/androidTest/java/com/epfl/drawyourpath/database/MockDatabaseTest.kt
@@ -11,7 +11,7 @@ import com.epfl.drawyourpath.chat.MessageContent
 import com.epfl.drawyourpath.community.Tournament
 import com.epfl.drawyourpath.path.Path
 import com.epfl.drawyourpath.path.Run
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
 import com.google.android.gms.maps.model.LatLng
 import org.junit.Assert.*
 import org.junit.Assert.assertEquals
