+++ app/src/androidTest/java/com/epfl/drawyourpath/database/MockNonWorkingDatabaseTest.kt
@@ -2,7 +2,7 @@ package com.epfl.drawyourpath.database
 
 import com.epfl.drawyourpath.path.Path
 import com.epfl.drawyourpath.path.Run
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.util.concurrent.CompletableFuture
