+++ app/src/androidTest/java/com/epfl/drawyourpath/userProfile/cache/UserModelCachedTest.kt
@@ -14,7 +14,7 @@ import com.epfl.drawyourpath.database.UserGoals
 import com.epfl.drawyourpath.path.Path
 import com.epfl.drawyourpath.path.Run
 import com.epfl.drawyourpath.userProfile.UserProfile
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
 import com.epfl.drawyourpath.utils.Utils
 import com.google.android.gms.maps.model.LatLng
 import org.junit.Assert.*
