+++ app/src/main/java/com/epfl/drawyourpath/database/Database.kt
@@ -4,7 +4,7 @@ import android.graphics.Bitmap
 import com.epfl.drawyourpath.chat.Message
 import com.epfl.drawyourpath.community.Tournament
 import com.epfl.drawyourpath.path.Run
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
 import java.util.concurrent.CompletableFuture
 
 data class UserGoals(
