+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/ChallengeFragment.kt
@@ -14,7 +14,8 @@ import com.epfl.drawyourpath.community.Tournament
 import com.epfl.drawyourpath.database.MockDatabase
 import com.epfl.drawyourpath.database.UserGoals
 import com.epfl.drawyourpath.userProfile.cache.UserModelCached
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoalViewAdapter
 
 /**
  * Fragment used to display different challenge about the user :
