+++ app/src/main/java/com/epfl/drawyourpath/userProfile/cache/UserDao.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.room.*
 import com.epfl.drawyourpath.path.cache.PointsEntity
 import com.epfl.drawyourpath.path.cache.RunEntity
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoalEntity
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoalEntity
 
 @Dao
 interface UserDao {
