+++ app/src/main/java/com/epfl/drawyourpath/userProfile/cache/UserDatabase.kt
@@ -5,8 +5,8 @@ import androidx.room.RoomDatabase
 import com.epfl.drawyourpath.path.cache.PointsEntity
 import com.epfl.drawyourpath.path.cache.RunDao
 import com.epfl.drawyourpath.path.cache.RunEntity
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoalDao
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoalEntity
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoalDao
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoalEntity
 
 @Database(entities = [UserEntity::class, DailyGoalEntity::class, RunEntity::class, PointsEntity::class], version = 5)
 abstract class UserDatabase : RoomDatabase() {
