+++ app/src/main/java/com/epfl/drawyourpath/userProfile/cache/UserModelCached.kt
@@ -12,8 +12,8 @@ import com.epfl.drawyourpath.database.*
 import com.epfl.drawyourpath.path.Run
 import com.epfl.drawyourpath.path.cache.RunEntity
 import com.epfl.drawyourpath.userProfile.UserProfile
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoal
-import com.epfl.drawyourpath.userProfile.dailygoal.DailyGoalEntity
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoal
+import com.epfl.drawyourpath.challenge.dailygoal.DailyGoalEntity
 import com.epfl.drawyourpath.utils.Utils
 import java.time.LocalDate
 import java.util.concurrent.CompletableFuture
